src/requirements.txt
# =========================
# Core (Django + DRF + JWT)
# =========================
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-filter==25.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
asgiref==3.9.1
sqlparse==0.5.3
python-dotenv==1.1.1
# (Escolha usar dotenv OU decouple; mantendo só dotenv)

# =========================
# Segurança / Login Hardening
# =========================
django-axes==8.0.0              # lockout de login
django-ratelimit==4.1.0         # rate-limit por rota (/login)
django-otp==1.6.1               # base pra MFA
django-two-factor-auth==1.17.0  # TOTP/2FA
pyotp==2.9.0
qrcode==7.4.2
python-json-logger==3.3.0
django-redis==6.0.0
redis==6.4.0

# =========================
# Ferramentas MPLS/NetOps
# =========================
paramiko==4.0.0      # (unifiquei: remova 3.5.1)
PyNaCl==1.5.0
openpyxl==3.1.5
PyYAML==6.0.2
inflection==0.5.1
requests==2.32.4

# =========================================
# OPCIONAL: Geo/GIS (ative só com libs do SO)
# =========================================
# GDAL==3.8.4
# shapely==2.1.1
# djangorestframework-gis==1.2.0
# geoip2==5.1.0
# maxminddb==2.8.2
# numpy==2.3.2

# =========================================
# OPCIONAL: LDAP
# =========================================
# django-auth-ldap==5.2.0
# python-ldap==3.4.4

# =========================================
# Dev / utilidades
# =========================================
django-extensions==4.1
